requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-api-key==2.1.0
dnspython==2.1.0
eventlet==0.33.0
greenlet==1.1.2
honcho==1.1.0
kombu==5.2.2
loguru==0.5.3
numpy==1.21.5
pandas==1.3.5
prompt-toolkit==3.0.24
psycopg2==2.9.2
python-dateutil==2.8.2
pytz==2021.3
redis==4.0.2
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
